Major refactoring in the meta model. I moved almost all functionality which isnt defined by the MetaModel interface into separate inner classes. One class for the internal dpf meta model, one class for the dpf model, and one class for the ecore specific stuff. Should be a bit more readable. Still needs cleaning up in the InternalModel methods.
The bug where we get wrong List type for getArrows and getNodes is still not fixed due to the fact that we cant reach the Node and Arrow     
object from the Specification. It might be possible to hardcode some method in ArrowType to do this. In the meantime, getters for each       
arrow type is created in the same manner as getters for nodes.